+++ app/integrations/woocommerce/templates/cart/cart-item.php
@@ -67,7 +67,7 @@ $product_permalink = apply_filters( 'woocommerce_cart_item_permalink', $_product
 				<?php
 				if ( 1 !== absint( $cart_item['quantity'] ) ) :
 					printf(
-						// Translators: %1$s cart item quantity. %2$s Cart item price.
+						/* translators: %1$s cart item quantity. %2$s Cart item price. */
 						__( '%1$s &times %2$s', 'bigbox' ),
 						absint( $cart_item['quantity'] ),
 						apply_filters( 'woocommerce_cart_item_price', WC()->cart->get_product_price( $_product ), $cart_item, $cart_item_key ) // PHPCS: XSS ok.