+++ app/integrations/woocommerce/templates/cart/cart-shipping.php
@@ -15,26 +15,35 @@
  * @see     https://docs.woocommerce.com/document/template-structure/
  * @author  WooThemes
  * @package WooCommerce/Templates
- * @version 3.2.0
+ * @version 3.5.0
  */
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
+
+$formatted_destination    = isset( $formatted_destination ) ? $formatted_destination : WC()->countries->get_formatted_address( $package['destination'], ', ' );
+$has_calculated_shipping  = ! empty( $has_calculated_shipping );
+$show_shipping_calculator = ! empty( $show_shipping_calculator );
+$calculator_text          = __( '(update address)', 'bigbox' );
+$note                     = false;
+$chosen_method_object     = false;
 ?>
 
 <div class="action-list__item">
 	<div id="package-name" class="action-list__item-label">
 		<?php echo wp_kses_post( $package_name ); ?>:
 	</div>
-	<div class="action-list__item-value">
-		<?php if ( ! is_checkout() && ! empty( $show_shipping_calculator ) ) : ?>
-		<a href="#" class="shipping-calculator-button"><?php echo esc_html__( 'Calculate shipping', 'bigbox' ); ?></a>
-		<?php endif; ?>
+	<div class="action-list__item-value woocommerce-shipping-address">
+		<?php
+		if ( $formatted_destination ) :
+			echo wp_kses( str_replace( '<br />', ', ', $formatted_destination ), [] );
+		endif;
+		?>
 	</div>
 </div>
 
-<?php if ( 1 < count( $available_methods ) ) : ?>
+<?php if ( $available_methods ) : ?>
 
 	<ul id="shipping_method" class="shipping-methods">
 		<?php foreach ( $available_methods as $method ) : ?>
@@ -49,56 +58,66 @@ if ( ! defined( 'ABSPATH' ) ) {
 						wc_cart_totals_shipping_method_label( $method ) // @codingStandardsIgnoreLine
 					);
 
+					// Track chosen method to display price later.
+					if ( $method->id === $chosen_method ) :
+						$chosen_method_object = $method;
+					endif;
+
 					do_action( 'woocommerce_after_shipping_rate', $method, $index );
 				?>
 			</li>
 		<?php endforeach; ?>
+
+		<?php if ( ! is_checkout() && $show_shipping_calculator ) : ?>
+			<li>
+				<button class="shipping-calculator-button button--text"><?php echo esc_html( $calculator_text ); ?></button>
+			</li>
+		<?php endif; ?>
 	</ul>
 
 <?php
-// Show single method.
-elseif ( 1 === count( $available_methods ) ) :
-	$method = current( $available_methods );
+elseif ( ! $formatted_destination && $available_methods ) :
+	$note = __( 'Enter your address to view shipping options.', 'bigbox' );
+elseif ( ! is_cart() ) :
+	$note = apply_filters( 'woocommerce_no_shipping_available_html', __( 'There are no shipping methods available. Please ensure that your address has been entered correctly, or contact us if you need any help.', 'bigbox' ) );
+else :
+	/* translators: %s shipping destination. */
+	$note= apply_filters( 'woocommerce_cart_no_shipping_available_html', sprintf( esc_html__( 'No shipping options were found for %s.', 'bigbox' ) . ' ', '<strong>' . esc_html( $formatted_destination ) . '</strong>' ) );
+endif;
+
+// Show note.
+if ( $note ) :
 ?>
 
-<div class="action-list__item action-list__item--inset">
-	<div class="action-list__item-label">
-		<?php echo wc_cart_totals_shipping_method_label( $method ); // @codingStandardsIgnoreLine ?>
-	</div>
-	<div class="action-list__item-value">
-		<?php echo bigbox_woocommerce_cart_shipping_method_price( $method ); // @codingStandardsIgnoreLine ?>
-	</div>
-</div>
+<p class="woocommerce-shipping-note">
+	<?php echo wp_kses_post( $note ); ?>
+
+	<?php if ( $show_shipping_calculator ) : ?>
+	<button class="shipping-calculator-button button--text"><?php echo esc_html( $calculator_text ); ?></button>
+	<?php endif; ?>
+</p>
 
 <?php
-	printf(
-		'<input type="hidden" name="shipping_method[%1$d]" data-index="%1$d" id="shipping_method_%1$d" value="%2$s" class="shipping_method" />',
-		esc_attr( $index ),
-		esc_attr( $method->id )
-	); // @codingStandardsIgnoreLine
-
-	do_action( 'woocommerce_after_shipping_rate', $method, $index );
-
-// No shipping.
-elseif ( WC()->customer->has_calculated_shipping() ) :
-	echo '<div class="shipping-note">';
-	echo wp_kses_post( apply_filters( is_cart() ? 'woocommerce_cart_no_shipping_available_html' : 'woocommerce_no_shipping_available_html', wpautop( __( 'There are no shipping methods available. Please ensure that your address has been entered correctly, or contact us if you need any help.', 'bigbox' ) ) ) );
-	echo '</div>';
-
-	// Unavailable.
-elseif ( ! is_cart() ) :
-	echo '<div class="shipping-note">';
-	echo wp_kses_post( wpautop( __( 'Enter your full address to see shipping costs.', 'bigbox' ) ) );
-	echo '</div>';
 endif;
 
 if ( $show_package_details ) :
-	echo '<div class="shipping-note">';
-	echo '<p class="woocommerce-shipping-contents"><small>' . esc_html( $package_details ) . '</small></p>';
-	echo '</div>';
+	echo '<p class="woocommerce-shipping-contents">' . esc_html( $package_details ) . '</p>';
+endif;
+
+if ( $show_shipping_calculator ) :
+	woocommerce_shipping_calculator( $calculator_text );
 endif;
 ?>
 
-<?php if ( ! empty( $show_shipping_calculator ) ) : ?>
-	<?php woocommerce_shipping_calculator(); ?>
+<?php if ( $chosen_method_object ) : ?>
+
+<div class="action-list__item">
+	<div class="action-list__item-label">
+		<?php esc_html_e( 'Shipping:', 'bigbox' ); ?>
+	</div>
+	<div class="action-list__item-value action-list__item-value--no-flex">
+		<?php echo wp_kses_post( bigbox_woocommerce_cart_shipping_method_price( $chosen_method_object ) ); ?>
+	</div>
+</div>
+
 <?php endif; ?>