+++ app/integrations/woocommerce/templates/loop/sale-flash.php
@@ -25,7 +25,7 @@ global $post, $product;
 if ( $product->is_on_sale() ) :
 	if ( 'simple' === $product->get_type() ) :
 		$percentage = round( ( $product->get_regular_price() - $product->get_sale_price() ) / $product->get_regular_price() * 100 );
-		// Translators: %1$s Sale savings amount.
+		/* translators: %1$s Sale savings amount. */
 		$flash = sprintf( __( 'Save %1$s%%', 'bigbox' ), $percentage );
 	else :
 		$flash = __( 'Sale', 'bigbox' );
@@ -34,7 +34,7 @@ if ( $product->is_on_sale() ) :
 
 <div class="product__sale">
 <?php
-// Translators: %1$s Sale price percentage.
+/* translators: %1$s Sale price percentage. */
 echo esc_html( apply_filters( 'woocommerce_sale_flash', $flash, $post, $product ) );
 ?>
 </div>